add dataclasses
